requirements.txt
fastapi
pydantic
requests
python-dotenv
supabase==2.5.1
PyJWT>=2.0.0